+++ ezplayer/lib/flutter_subtiltle_wrapper/data/models/subtitles.dart
@@ -1,0 +1,9 @@
+import 'subtitle.dart';
+
+class Subtitles {
+  final List<Subtitle> subtitles;
+
+  Subtitles({
+    required this.subtitles,
+  });
+}